feat: Implement AppContext and related hooks for global state management
- Created AppContext to manage application-wide state including clients, team members, workflows, meetings, kanban columns, tasks, and dashboard stats.
- Developed useAppContext hook for easy access to AppContext.
- Added useClientActions, useMeetingActions, useTaskActions, useTeamActions, and useWorkflowActions hooks for CRUD operations on respective entities.
- Implemented ClientsPage, DashboardPage, KanbanPage, MeetingsPage, TeamPage, and WorkflowsPage components to utilize the new context and hooks.
- Integrated modals for creating and editing clients, meetings, team members, tasks, and workflows.
- Enhanced error handling and loading states during API interactions.
