+++ index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <link rel="icon" href="/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>React Flow starter (Vite + TS)</title>
+    <title>Benders Corp Workflow</title>
   </head>
   <body>
     <div id="root"></div>
