+++ package.json
@@ -9,13 +9,15 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@types/react-router-dom": "^5.3.3",
     "@xyflow/react": "^12.5.1",
+    "date-fns": "^2.30.0",
+    "lucide-react": "^0.294.0",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0",
     "react-dnd": "^16.0.1",
     "react-dnd-html5-backend": "^16.0.1",
-    "lucide-react": "^0.294.0",
-    "date-fns": "^2.30.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^7.6.2",
     "uuid": "^9.0.1"
   },
   "license": "MIT",
